requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
anthropic>=0.30.0
openai>=1.0.0
tavily-python>=0.3.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
